+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Python lib: Support for clearing a retained setting
 * Python CLI: get() support
+* `TreeKey::iter_indices()` and `iter_indices_unchecked()`
 
 ## [0.9.0](https://github.com/quartiq/miniconf/compare/v0.8.0...v0.9.0)